Extract SongImporter concerns for audio recognition, track finding, air play creation, and artist updating
